requirements.txt
pbr>=0.5.21,<1.0
eventlet>=0.13.0
greenlet>=0.3.2
httplib2
iso8601>=0.1.4
argparse
lxml>=2.3
netaddr
six>=1.4.1
PasteDeploy>=1.5.0
Routes>=1.12.3
WebOb>=1.2.3,<1.3
PyYAML>=3.1.0
paramiko>=1.8.0
PyGithub>=1.14.0
Babel>=1.3
oslo.config>=1.2.0
redis>=2.7.6
pylibmc>=1.2.3
requests>=1.2.3
SQLAlchemy>=0.7.8,<=0.7.99
mysql-python
